chore: fixed some minor bugs

+++ nft_staking_q3/programs/nft_staking_q3/src/state/stake_config.rs
@@ -5,7 +5,7 @@ use anchor_lang::prelude::*;
 pub struct StakeConfig {
     pub points_per_stake: u8,
     pub max_staked: u8,
-    pub freeze_period: u32,
+    pub freeze_period: i64,
     pub reward_bump: u8,
     pub bump: u8,
 }
